BackEnd/requirements.txt
blinker==1.6.2
certifi==2022.9.24
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
coverage==6.5.0
Flask==2.3.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.0.2
Werkzeug==2.3.4
